Clean up obsolete behavior page and components (#2986)
* Remove old behavior page code

* Refactor behavior

* Remove unnecessary filter now there's single value support

* Clean up obsolete feature flag

* Clean up behavior threshold configuration
