+++ packages/app/src/domain/behavior/behavior-choropleths-tile.tsx
@@ -1,5 +1,4 @@
 import {
-  ChoroplethThresholdsValue,
   RegionsBehavior,
   SafetyRegionProperties,
 } from '@corona-dashboard/common';
@@ -10,20 +9,22 @@ import { regionThresholds } from '~/components/choropleth/region-thresholds';
 import { SafetyRegionChoropleth } from '~/components/choropleth/safety-region-choropleth';
 import { BehaviorTooltip } from '~/components/choropleth/tooltips/region/behavior-tooltip';
 import { ErrorBoundary } from '~/components/error-boundary';
-import { Select } from '~/components/select';
 import { Tile } from '~/components/tile';
 import { Heading, Text } from '~/components/typography';
 import { useIntl } from '~/intl';
 import { colors } from '~/style/theme';
 import { useBreakpoints } from '~/utils/use-breakpoints';
 import { useReverseRouter } from '~/utils/use-reverse-router';
-import { BehaviorIdentifier, behaviorIdentifiers } from '../behavior-types';
-import { BehaviorIcon } from '../components/behavior-icon';
+import { SelectBehavior } from './components/select-behavior';
+import {
+  BehaviorIdentifier,
+  behaviorIdentifiers,
+} from './logic/behavior-types';
 
 interface BehaviorChoroplethsTileProps {
   title: string;
   description: string;
-  data: RegionsBehavior[];
+  data: { behavior: RegionsBehavior[] };
   currentId: BehaviorIdentifier;
   setCurrentId: React.Dispatch<React.SetStateAction<BehaviorIdentifier>>;
 }
@@ -36,7 +37,7 @@ export function BehaviorChoroplethsTile({
   setCurrentId,
 }: BehaviorChoroplethsTileProps) {
   const { siteText } = useIntl();
-  const firstRegionData = data[0];
+  const firstRegionData = data.behavior[0];
 
   // Find all the keys that don't exist on VR level but do on NL
   const keysWithoutData = behaviorIdentifiers.filter(
@@ -56,14 +57,6 @@ export function BehaviorChoroplethsTile({
 
   keysWithoutData.push(...(idsThatContainNull as BehaviorIdentifier[]));
 
-  const behaviorIndentifiersData = behaviorIdentifiers.map((id) => {
-    const label = siteText.gedrag_onderwerpen[id];
-    return {
-      label,
-      value: id,
-    };
-  });
-
   return (
     <Tile>
       <Heading level={3}>{title}</Heading>
@@ -72,31 +65,24 @@ export function BehaviorChoroplethsTile({
       </Box>
 
       <Box mb={4}>
-        <Select
-          value={currentId}
-          onChange={setCurrentId}
-          options={behaviorIndentifiersData}
-          icon={<BehaviorIcon name={currentId} size={20} />}
-        />
+        <SelectBehavior value={currentId} onChange={setCurrentId} />
       </Box>
 
       <Box display="flex" flexWrap="wrap" spacing={{ _: 4, lg: undefined }}>
         <ChoroplethBlock
           title={siteText.nl_gedrag.verdeling_in_nederland.compliance_title}
-          data={{ behavior_compliance: data }}
+          data={data}
           behaviorType="compliance"
           currentId={currentId}
           keysWithoutData={keysWithoutData}
-          thresholds={regionThresholds.behavior_compliance}
         />
 
         <ChoroplethBlock
           title={siteText.nl_gedrag.verdeling_in_nederland.support_title}
-          data={{ behavior_support: data }}
+          data={data}
           behaviorType="support"
           currentId={currentId}
           keysWithoutData={keysWithoutData}
-          thresholds={regionThresholds.behavior_support}
         />
       </Box>
     </Tile>
@@ -104,12 +90,11 @@ export function BehaviorChoroplethsTile({
 }
 
 interface ChoroplethBlockProps {
-  data: { [key: string]: RegionsBehavior[] };
+  data: { behavior: RegionsBehavior[] };
   keysWithoutData: BehaviorIdentifier[];
   behaviorType: 'compliance' | 'support';
   currentId: BehaviorIdentifier;
   title: string;
-  thresholds: ChoroplethThresholdsValue[];
 }
 
 function ChoroplethBlock({
@@ -118,13 +103,13 @@ function ChoroplethBlock({
   behaviorType,
   currentId,
   title,
-  thresholds,
 }: ChoroplethBlockProps) {
   const { siteText } = useIntl();
   const reverseRouter = useReverseRouter();
   const breakpoints = useBreakpoints();
 
   const isSmallScreen = breakpoints.sm;
+  const metricProperty = `${currentId}_${behaviorType}` as const;
 
   return (
     <Box width={{ _: '100%', lg: '50%' }}>
@@ -150,10 +135,10 @@ function ChoroplethBlock({
         )}
         <ErrorBoundary>
           <SafetyRegionChoropleth
-            data={data as { behavior: RegionsBehavior[] }}
+            data={data}
             getLink={reverseRouter.vr.gedrag}
-            metricName={`behavior_${behaviorType}` as 'behavior'}
-            metricProperty={`${currentId}_${behaviorType}`}
+            metricName="behavior"
+            metricProperty={metricProperty}
             minHeight={!isSmallScreen ? 350 : 400}
             noDataFillColor={colors.page}
             tooltipContent={(
@@ -188,7 +173,7 @@ function ChoroplethBlock({
         maxWidth={300}
       >
         <ChoroplethLegenda
-          thresholds={thresholds}
+          thresholds={regionThresholds.behavior[metricProperty]}
           title={siteText.gedrag_common.basisregels.header_percentage}
         />
       </Box>
