+++ packages/app/src/pages/landelijk/gedrag.tsx
@@ -1,22 +1,23 @@
+import { useMemo, useRef, useState } from 'react';
 import Gedrag from '~/assets/gedrag.svg';
 import { ArticleStrip } from '~/components/article-strip';
 import { ArticleSummary } from '~/components/article-teaser';
 import { ContentHeader } from '~/components/content-header';
-import { KpiTile } from '~/components/kpi-tile';
-import { KpiValue } from '~/components/kpi-value';
+import { Markdown } from '~/components/markdown';
 import { Tile } from '~/components/tile';
 import { TileList } from '~/components/tile-list';
 import { TwoKpiSection } from '~/components/two-kpi-section';
 import { Heading, Text } from '~/components/typography';
-import { BehaviorChoroplethTile } from '~/domain/behavior/behavior-choropleth-tile';
+import { BehaviorChoroplethsTile } from '~/domain/behavior/behavior-choropleths-tile';
 import { BehaviorLineChartTile } from '~/domain/behavior/behavior-line-chart-tile';
+import { BehaviorPerAgeGroup } from '~/domain/behavior/behavior-per-age-group-tile';
 import { BehaviorTableTile } from '~/domain/behavior/behavior-table-tile';
-import { BehaviorPageNational } from '~/domain/behavior/redesign/_behavior-page-national';
 import { MoreInformation } from '~/domain/behavior/components/more-information';
+import { BehaviorIdentifier } from '~/domain/behavior/logic/behavior-types';
+import { useBehaviorLookupKeys } from '~/domain/behavior/logic/use-behavior-lookup-keys';
 import { Layout } from '~/domain/layout/layout';
 import { NationalLayout } from '~/domain/layout/national-layout';
 import { useIntl } from '~/intl';
-import { useFeature } from '~/lib/features';
 import { createPageArticlesQuery } from '~/queries/create-page-articles-query';
 import {
   createGetStaticProps,
@@ -28,6 +29,7 @@ import {
   getLastGeneratedDate,
   selectNlPageMetricData,
 } from '~/static-props/get-data';
+import { replaceVariablesInText } from '~/utils/replace-variables-in-text';
 
 export const getStaticProps = createGetStaticProps(
   getLastGeneratedDate,
@@ -43,94 +45,168 @@ export const getStaticProps = createGetStaticProps(
   })
 );
 
-const BehaviorPage = (props: StaticProps<typeof getStaticProps>) => {
-  const { siteText } = useIntl();
-
+export default function BehaviorPage(
+  props: StaticProps<typeof getStaticProps>
+) {
   const { selectedNlData: data, choropleth, content, lastGenerated } = props;
   const behaviorLastValue = data.behavior.last_value;
-  const { nl_gedrag } = siteText;
+
+  const intl = useIntl();
+  const { nl_gedrag } = intl.siteText;
 
   const metadata = {
-    ...siteText.nationaal_metadata,
+    ...intl.siteText.nationaal_metadata,
     title: nl_gedrag.metadata.title,
     description: nl_gedrag.metadata.description,
   };
 
-  const featureBehaviorPage = useFeature('behaviorPage');
+  const [currentId, setCurrentId] = useState<BehaviorIdentifier>('wash_hands');
+  const scrollToRef = useRef<HTMLDivElement>(null);
+
+  const behaviorLookupKeys = useBehaviorLookupKeys();
+
+  const { highestCompliance, highestSupport } = useMemo(() => {
+    const list = behaviorLookupKeys.map((x) => ({
+      description: x.description,
+      compliancePercentage: behaviorLastValue[x.complianceKey] as number,
+      supportPercentage: behaviorLastValue[x.supportKey] as number,
+    }));
+
+    const highestCompliance = list.sort(
+      (a, b) => (b.compliancePercentage ?? 0) - (a.compliancePercentage ?? 0)
+    )[0];
+
+    const highestSupport = list.sort(
+      (a, b) => (b.supportPercentage ?? 0) - (a.supportPercentage ?? 0)
+    )[0];
+
+    return { highestCompliance, highestSupport };
+  }, [behaviorLastValue, behaviorLookupKeys]);
 
   return (
     <Layout {...metadata} lastGenerated={lastGenerated}>
       <NationalLayout data={data} lastGenerated={lastGenerated}>
-        {featureBehaviorPage.isEnabled ? (
-          <BehaviorPageNational
-            data={data}
-            content={content}
-            behaviorData={choropleth.vr.behavior}
+        <TileList>
+          <ContentHeader
+            category={intl.siteText.nationaal_layout.headings.gedrag}
+            title={nl_gedrag.pagina.titel}
+            icon={<Gedrag />}
+            subtitle={nl_gedrag.pagina.toelichting}
+            metadata={{
+              datumsText: nl_gedrag.datums,
+              dateOrRange: {
+                start: behaviorLastValue.date_start_unix,
+                end: behaviorLastValue.date_end_unix,
+              },
+              dateOfInsertionUnix: behaviorLastValue.date_of_insertion_unix,
+              dataSources: [nl_gedrag.bronnen.rivm],
+            }}
+            reference={nl_gedrag.reference}
           />
-        ) : (
-          <TileList>
-            <ContentHeader
-              category={siteText.nationaal_layout.headings.gedrag}
-              title={nl_gedrag.pagina.titel}
-              icon={<Gedrag />}
-              subtitle={nl_gedrag.pagina.toelichting}
-              metadata={{
-                datumsText: nl_gedrag.datums,
-                dateOrRange: {
-                  start: behaviorLastValue.date_start_unix,
-                  end: behaviorLastValue.date_end_unix,
-                },
-                dateOfInsertionUnix: behaviorLastValue.date_of_insertion_unix,
-                dataSources: [nl_gedrag.bronnen.rivm],
-              }}
-              reference={nl_gedrag.reference}
+
+          <TwoKpiSection>
+            <Tile>
+              <Heading level={3}>{nl_gedrag.onderzoek_uitleg.titel}</Heading>
+              <Text>{nl_gedrag.onderzoek_uitleg.toelichting}</Text>
+            </Tile>
+            <Tile>
+              <Heading level={3}>
+                {nl_gedrag.kpi_recente_inzichten.titel}
+              </Heading>
+
+              <Markdown
+                content={replaceVariablesInText(
+                  nl_gedrag.kpi_recente_inzichten.tekst,
+                  {
+                    number_of_participants: intl.formatNumber(
+                      behaviorLastValue.number_of_participants
+                    ),
+                    date_start: intl.formatDateFromSeconds(
+                      behaviorLastValue.date_start_unix
+                    ),
+                    date_end: intl.formatDateFromSeconds(
+                      behaviorLastValue.date_end_unix
+                    ),
+
+                    highest_compliance_description:
+                      highestCompliance.description,
+                    highest_compliance_compliance_percentage:
+                      intl.formatPercentage(
+                        highestCompliance.compliancePercentage
+                      ),
+                    highest_compliance_support_percentage:
+                      intl.formatPercentage(
+                        highestCompliance.supportPercentage
+                      ),
+
+                    highest_support_description: highestSupport.description,
+                    highest_support_compliance_percentage:
+                      intl.formatPercentage(
+                        highestSupport.compliancePercentage
+                      ),
+                    highest_support_support_percentage: intl.formatPercentage(
+                      highestSupport.supportPercentage
+                    ),
+                  }
+                )}
+              />
+            </Tile>
+          </TwoKpiSection>
+
+          <ArticleStrip articles={content.articles} />
+
+          <BehaviorTableTile
+            title={nl_gedrag.basisregels.title}
+            description={nl_gedrag.basisregels.description}
+            complianceExplanation={nl_gedrag.basisregels.volgen_beschrijving}
+            supportExplanation={nl_gedrag.basisregels.steunen_beschrijving}
+            value={behaviorLastValue}
+            annotation={nl_gedrag.basisregels.annotatie}
+            setCurrentId={setCurrentId}
+            scrollRef={scrollToRef}
+          />
+
+          <span ref={scrollToRef} />
+          <BehaviorLineChartTile
+            values={data.behavior.values}
+            metadata={{
+              date: [
+                behaviorLastValue.date_start_unix,
+                behaviorLastValue.date_end_unix,
+              ],
+              source: nl_gedrag.bronnen.rivm,
+            }}
+            currentId={currentId}
+            setCurrentId={setCurrentId}
+          />
+
+          <BehaviorChoroplethsTile
+            title={nl_gedrag.verdeling_in_nederland.titel}
+            description={nl_gedrag.verdeling_in_nederland.description}
+            data={choropleth.vr}
+            currentId={currentId}
+            setCurrentId={setCurrentId}
+          />
+
+          {data.behavior_per_age_group && (
+            <BehaviorPerAgeGroup
+              title={nl_gedrag.tabel_per_leeftijdsgroep.title}
+              description={nl_gedrag.tabel_per_leeftijdsgroep.description}
+              complianceExplanation={
+                nl_gedrag.tabel_per_leeftijdsgroep.explanation.compliance
+              }
+              supportExplanation={
+                nl_gedrag.tabel_per_leeftijdsgroep.explanation.support
+              }
+              data={data.behavior_per_age_group}
+              currentId={currentId}
+              setCurrentId={setCurrentId}
             />
+          )}
 
-            <ArticleStrip articles={content.articles} />
-
-            <TwoKpiSection>
-              <Tile height="100%">
-                <Heading level={3}>{nl_gedrag.onderzoek_uitleg.titel}</Heading>
-                <Text>{nl_gedrag.onderzoek_uitleg.toelichting}</Text>
-              </Tile>
-
-              <KpiTile
-                title={nl_gedrag.kpi.aantal_respondenten.titel}
-                metadata={{
-                  source: nl_gedrag.kpi.aantal_respondenten.bron,
-                  date: [
-                    behaviorLastValue.date_start_unix,
-                    behaviorLastValue.date_end_unix,
-                  ],
-                }}
-              >
-                <KpiValue absolute={behaviorLastValue.number_of_participants} />
-                <Text>{nl_gedrag.kpi.aantal_respondenten.toelichting}</Text>
-              </KpiTile>
-            </TwoKpiSection>
-
-            <BehaviorTableTile
-              behavior={behaviorLastValue}
-              title={nl_gedrag.basisregels.title}
-              introduction={nl_gedrag.basisregels.intro}
-              footer={nl_gedrag.basisregels.voetnoot}
-              footerAsterisk={nl_gedrag.basisregels.voetnoot_asterisk}
-            />
-
-            <BehaviorLineChartTile
-              values={data.behavior.values}
-              title={nl_gedrag.basisregels_over_tijd.title}
-              introduction={nl_gedrag.basisregels_over_tijd.intro}
-            />
-
-            <BehaviorChoroplethTile data={choropleth.vr} />
-
-            <MoreInformation />
-          </TileList>
-        )}
+          <MoreInformation />
+        </TileList>
       </NationalLayout>
     </Layout>
   );
-};
-
-export default BehaviorPage;
+}
