+++ packages/app/src/pages/veiligheidsregio/[code]/gedrag.tsx
@@ -1,21 +1,19 @@
+import { useRef, useState } from 'react';
 import Gedrag from '~/assets/gedrag.svg';
 import { ArticleStrip } from '~/components/article-strip';
 import { ArticleSummary } from '~/components/article-teaser';
 import { ContentHeader } from '~/components/content-header';
-import { KpiTile } from '~/components/kpi-tile';
-import { KpiValue } from '~/components/kpi-value';
 import { Tile } from '~/components/tile';
 import { TileList } from '~/components/tile-list';
 import { TwoKpiSection } from '~/components/two-kpi-section';
-import { Heading, Text } from '~/components/typography';
+import { Heading, InlineText, Text } from '~/components/typography';
 import { BehaviorLineChartTile } from '~/domain/behavior/behavior-line-chart-tile';
-import { BehaviorPageSafetyRegion } from '~/domain/behavior/redesign/_behavior-page-safety-region';
 import { BehaviorTableTile } from '~/domain/behavior/behavior-table-tile';
 import { MoreInformation } from '~/domain/behavior/components/more-information';
+import { BehaviorIdentifier } from '~/domain/behavior/logic/behavior-types';
 import { Layout } from '~/domain/layout/layout';
 import { SafetyRegionLayout } from '~/domain/layout/safety-region-layout';
 import { useIntl } from '~/intl';
-import { useFeature } from '~/lib/features';
 import { createPageArticlesQuery } from '~/queries/create-page-articles-query';
 import {
   createGetStaticProps,
@@ -26,6 +24,8 @@ import {
   getLastGeneratedDate,
   selectVrPageMetricData,
 } from '~/static-props/get-data';
+import { replaceComponentsInText } from '~/utils/replace-components-in-text';
+
 export { getStaticPaths } from '~/static-paths/vr';
 
 export const getStaticProps = createGetStaticProps(
@@ -39,16 +39,18 @@ export const getStaticProps = createGetStaticProps(
   })
 );
 
-const BehaviorPage = (props: StaticProps<typeof getStaticProps>) => {
+export default function BehaviorPageSafetyRegion(
+  props: StaticProps<typeof getStaticProps>
+) {
   const {
     lastGenerated,
     content,
     selectedVrData: data,
     safetyRegionName,
   } = props;
-  const behaviorData = data.behavior;
 
-  const { siteText } = useIntl();
+  const { siteText, formatDateFromSeconds, formatNumber } = useIntl();
+
   const text = siteText.regionaal_gedrag;
 
   const metadata = {
@@ -57,7 +59,11 @@ const BehaviorPage = (props: StaticProps<typeof getStaticProps>) => {
     description: text.metadata.description,
   };
 
-  const featureBehaviorPage = useFeature('behaviorPage');
+  const { regionaal_gedrag } = siteText;
+  const behaviorLastValue = data.behavior.last_value;
+
+  const [currentId, setCurrentId] = useState<BehaviorIdentifier>('wash_hands');
+  const scrollToRef = useRef<HTMLDivElement>(null);
 
   return (
     <Layout {...metadata} lastGenerated={lastGenerated}>
@@ -66,73 +72,94 @@ const BehaviorPage = (props: StaticProps<typeof getStaticProps>) => {
         safetyRegionName={safetyRegionName}
         lastGenerated={lastGenerated}
       >
-        {featureBehaviorPage.isEnabled ? (
-          <BehaviorPageSafetyRegion data={data} content={content} />
-        ) : (
-          <TileList>
-            <ContentHeader
-              category={siteText.nationaal_layout.headings.gedrag}
-              title={text.pagina.titel}
-              icon={<Gedrag />}
-              subtitle={text.pagina.toelichting}
-              metadata={{
-                datumsText: text.datums,
-                dateOrRange: {
-                  start: behaviorData.last_value.date_start_unix,
-                  end: behaviorData.last_value.date_end_unix,
-                },
-                dateOfInsertionUnix:
-                  behaviorData.last_value.date_of_insertion_unix,
-                dataSources: [text.bronnen.rivm],
-              }}
-              reference={text.reference}
-            />
+        <TileList>
+          <ContentHeader
+            category={siteText.nationaal_layout.headings.gedrag}
+            title={regionaal_gedrag.pagina.titel}
+            icon={<Gedrag />}
+            subtitle={regionaal_gedrag.pagina.toelichting}
+            metadata={{
+              datumsText: regionaal_gedrag.datums,
+              dateOrRange: {
+                start: behaviorLastValue.date_start_unix,
+                end: behaviorLastValue.date_end_unix,
+              },
+              dateOfInsertionUnix: behaviorLastValue.date_of_insertion_unix,
+              dataSources: [regionaal_gedrag.bronnen.rivm],
+            }}
+            reference={regionaal_gedrag.reference}
+          />
 
-            <ArticleStrip articles={content.articles} />
+          <TwoKpiSection>
+            <Tile>
+              <Heading level={3}>
+                {regionaal_gedrag.onderzoek_uitleg.titel}
+              </Heading>
+              <Text>{regionaal_gedrag.onderzoek_uitleg.toelichting}</Text>
+            </Tile>
+            <Tile height="100%">
+              <Heading level={3}>{regionaal_gedrag.kpi.titel}</Heading>
+              <Text>
+                {replaceComponentsInText(
+                  regionaal_gedrag.kpi.deelgenomen_mensen,
+                  {
+                    number_of_participants: (
+                      <InlineText fontWeight="bold">
+                        {formatNumber(behaviorLastValue.number_of_participants)}
+                      </InlineText>
+                    ),
+                    date_start: (
+                      <InlineText>
+                        {formatDateFromSeconds(
+                          behaviorLastValue.date_start_unix
+                        )}
+                      </InlineText>
+                    ),
+                    date_end: (
+                      <InlineText>
+                        {formatDateFromSeconds(behaviorLastValue.date_end_unix)}
+                      </InlineText>
+                    ),
+                  }
+                )}
+              </Text>
+            </Tile>
+          </TwoKpiSection>
 
-            <TwoKpiSection>
-              <Tile height="100%">
-                <Heading level={3}>{text.onderzoek_uitleg.titel}</Heading>
-                <Text>{text.onderzoek_uitleg.toelichting}</Text>
-              </Tile>
+          <ArticleStrip articles={content.articles} />
 
-              <KpiTile
-                title={text.kpi.aantal_respondenten.titel}
-                metadata={{
-                  source: text.kpi.aantal_respondenten.bron,
-                  date: [
-                    behaviorData.last_value.date_start_unix,
-                    behaviorData.last_value.date_end_unix,
-                  ],
-                }}
-              >
-                <KpiValue
-                  absolute={behaviorData.last_value.number_of_participants}
-                />
-                <Text>{text.kpi.aantal_respondenten.toelichting}</Text>
-              </KpiTile>
-            </TwoKpiSection>
+          <BehaviorTableTile
+            title={regionaal_gedrag.basisregels.title}
+            description={regionaal_gedrag.basisregels.description}
+            complianceExplanation={
+              regionaal_gedrag.basisregels.volgen_beschrijving
+            }
+            supportExplanation={
+              regionaal_gedrag.basisregels.steunen_beschrijving
+            }
+            value={behaviorLastValue}
+            annotation={regionaal_gedrag.basisregels.annotatie}
+            setCurrentId={setCurrentId}
+            scrollRef={scrollToRef}
+          />
 
-            <BehaviorTableTile
-              behavior={behaviorData.last_value}
-              title={text.basisregels.title}
-              introduction={text.basisregels.intro}
-              footer={text.basisregels.voetnoot}
-              footerAsterisk={text.basisregels.voetnoot_asterisk}
-            />
+          <span ref={scrollToRef} />
+          <BehaviorLineChartTile
+            values={data.behavior.values}
+            metadata={{
+              date: [
+                behaviorLastValue.date_start_unix,
+                behaviorLastValue.date_end_unix,
+              ],
+              source: regionaal_gedrag.bronnen.rivm,
+            }}
+            currentId={currentId}
+            setCurrentId={setCurrentId}
+          />
 
-            <BehaviorLineChartTile
-              title={text.basisregels_over_tijd.title}
-              introduction={text.basisregels_over_tijd.intro}
-              values={behaviorData.values}
-            />
-
-            <MoreInformation />
-          </TileList>
-        )}
+          <MoreInformation />
+        </TileList>
       </SafetyRegionLayout>
     </Layout>
   );
-};
-
-export default BehaviorPage;
+}
